+++ src/types.ts
@@ -10,4 +10,7 @@ export {
   type FactorySeedColumnsInput,
   type FactorySeedFunctions,
   type FactorySeedGenerator,
+  type FactoryTraitInput,
+  type FactoryTraitRegistry,
+  type FactoryTraitsInput,
 } from "./internal/types";